admin_apps/pyproject.toml
[tool.poetry]
name = "admin_apps"
version = "0.1.0"
description = "Generate a Semantic Model from your Snowflake tables"
authors = ["Jonathan Hilgart <jonathan.hilgart@snowflake.com>", "Nipun Sehrawat <nipun.sehrawat@snowflake.com>", "Renee Huang <renee.huang@snowflake.com>", "Nicole Limtiaco <nicole.limtiaco@snowflake.com>"]
license = "Apache Software License; BSD License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0.0"
numpy = "^1.24.4"
pandas = "^2.0.1"
loguru = "^0.7.2"
snowflake-connector-python = {extras = ["pandas"], version = "^3.8.1"}
protobuf = "^4.25.3"
pydantic = "^2.6.4"
PyYAML = "^6.0.1"
"ruamel.yaml" = "^0.18.6"
jsonargparse = "^4.27.7"
tqdm = "^4.66.2"
pyarrow = ">=10.0.1,<10.1.0"
sqlglot = "^23.10.0"
strictyaml = "^1.7.3"
streamlit-nightly = "^1.35.1.dev20240609"
streamlit-extras = "0.4.3"
streamlit-monaco = "0.1.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.1.1"
types-pyyaml = "^6.0.12.20240311"
types-protobuf = "^4.24.0.20240311"
pip-licenses = "^4.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
